requirements.txt
Django==2.1.3
djangorestframework==3.9.0
pytz==2018.7
djangorestframework-simplejwt==3.2.3
